unreal: libpng16 NEW + x64 version
git-svn-id: svn://localhost/local-pentevo@1081 c6c76956-fad2-4150-a3a9-ea0a427205ba

+++ tools/unreal_fix/0.39.0/fix_labels/z80/tables.h
@@ -92,7 +92,7 @@ extern const TMakeSbc MakeSbc;
 Z80INLINE void and8(Z80 *cpu, unsigned char src)
 {
    cpu->a &= src;
-   cpu->f = log_f[cpu->a] | HF;
+   cpu->f = log_f[cpu->a] | (u8)HF;
 }
 
 Z80INLINE void or8(Z80 *cpu, unsigned char src)
